HDRenderPipeline: Rename HDRenderPipeline to HDRenderPipelineAsset and split in dedicated file
- Rename HDRenderPipeline to HDRenderPipelineAsset
- Move new renamed class in its own file
- Rename HDRenderPipelineInstance to HDRenderPipeline
