+++ Assets/ScriptableRenderPipeline/HDRenderPipeline/HDRenderPipelineAsset.asset.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: e185fecca3c73cd47a09f1092663ef32
-timeCreated: 1487689448
+guid: fa7ff29edd844404fb648e3eb1134f14
+timeCreated: 1496140270
 licenseType: Pro
 NativeFormatImporter:
   mainObjectFileID: 11400000
