+++ Assets/ScriptableRenderPipeline/HDRenderPipeline/HDRenderPipelineAsset.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0cf1dab834d4ec34195b920ea7bbf9ec
+timeCreated: 1496139874
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
